test/perf/pyproject.toml
[tool.poetry]
name = "perf"
version = "0.1.0"
description = ""
authors = ["Matthew McRaven <matthew.mcraven@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.7"
pygit2 = "==1.16.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
